+++ src/main/java/com/barco/model/dto/request/ControlRequest.java
@@ -35,6 +35,8 @@ public class ControlRequest extends RequestFilter {
     private ACTION action; // delete|insert
     private List<Long> sectionId;
     private List<Long> controlLinkSection;
+    private String disabledPattern;
+    private String visiblePattern;
     private Long controlOrder;
 
     public ControlRequest() {
@@ -206,6 +208,22 @@ public class ControlRequest extends RequestFilter {
 
     public void setControlLinkSection(List<Long> controlLinkSection) {
         this.controlLinkSection = controlLinkSection;
+    }
+
+    public String getDisabledPattern() {
+        return disabledPattern;
+    }
+
+    public void setDisabledPattern(String disabledPattern) {
+        this.disabledPattern = disabledPattern;
+    }
+
+    public String getVisiblePattern() {
+        return visiblePattern;
+    }
+
+    public void setVisiblePattern(String visiblePattern) {
+        this.visiblePattern = visiblePattern;
     }
 
     public Long getControlOrder() {
