+++ src/main/java/com/barco/model/dto/request/SectionRequest.java
@@ -23,6 +23,8 @@ public class SectionRequest extends RequestFilter {
     private List<Long> controlId;
     private List<Long> sectionLinkControl;
     private Long controlOrder;
+    private String disabledPattern;
+    private String visiblePattern;
     //
     private List<Long> formId;
     private List<Long> sectionLinkForm;
@@ -97,6 +99,22 @@ public class SectionRequest extends RequestFilter {
 
     public void setControlOrder(Long controlOrder) {
         this.controlOrder = controlOrder;
+    }
+
+    public String getDisabledPattern() {
+        return disabledPattern;
+    }
+
+    public void setDisabledPattern(String disabledPattern) {
+        this.disabledPattern = disabledPattern;
+    }
+
+    public String getVisiblePattern() {
+        return visiblePattern;
+    }
+
+    public void setVisiblePattern(String visiblePattern) {
+        this.visiblePattern = visiblePattern;
     }
 
     public List<Long> getFormId() {
