+++ src/main/java/com/barco/model/dto/response/SectionLinkControlResponse.java
@@ -15,6 +15,8 @@ public class SectionLinkControlResponse extends BaseEntity {
     private String controlName;
     private GLookup fieldType;
     private Long controlOrder;
+    private String disabledPattern;
+    private String visiblePattern;
     private Boolean linkedSection;
     private Long linkControlId;
 
@@ -43,6 +45,22 @@ public class SectionLinkControlResponse extends BaseEntity {
 
     public void setControlOrder(Long controlOrder) {
         this.controlOrder = controlOrder;
+    }
+
+    public String getDisabledPattern() {
+        return disabledPattern;
+    }
+
+    public void setDisabledPattern(String disabledPattern) {
+        this.disabledPattern = disabledPattern;
+    }
+
+    public String getVisiblePattern() {
+        return visiblePattern;
+    }
+
+    public void setVisiblePattern(String visiblePattern) {
+        this.visiblePattern = visiblePattern;
     }
 
     public Boolean getLinkedSection() {
