--- src/main/java/com/barco/model/repository/GenControlInteractionsRepository.java
@@ -1,13 +1,0 @@
-package com.barco.model.repository;
-
-import com.barco.model.pojo.GenControlInteractions;
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-/**
- * @author Nabeel Ahmed
- */
-@Repository
-public interface GenControlInteractionsRepository extends CrudRepository<GenControlInteractions, Long> {
-
-}
